Switch away from header-only style, which was causing link errors
