+++ packager/media/CMakeLists.txt
@@ -1,0 +1,8 @@
+# Copyright 2022 Google Inc. All rights reserved.
+#
+# Use of this source code is governed by a BSD-style
+# license that can be found in the LICENSE file or at
+# https://developers.google.com/open-source/licenses/bsd
+
+# Subdirectories with their own CMakeLists.txt, all of whose targets are built.
+add_subdirectory(test)
