+++ packager/media/test/test_data_util.h
@@ -9,44 +9,19 @@
 
 #include <filesystem>
 #include <string>
-
-#include "glog/logging.h"
+#include <vector>
 
 namespace shaka {
 namespace media {
 
 // Returns a file path for a file in the media/test/data directory.
-std::filesystem::path GetTestDataFilePath(const std::string& name) {
-  std::filesystem::path header_path(__FILE__);
-  return header_path.parent_path() / "data" / name;
-}
+std::filesystem::path GetTestDataFilePath(const std::string& name);
 
 // Returns a file path for a file in the media/app/test/testdata directory.
-std::filesystem::path GetAppTestDataFilePath(const std::string& name) {
-  std::filesystem::path header_path(__FILE__);
-  return header_path.parent_path().parent_path() / "app" / "test" / "testdata" /
-         name;
-}
+std::filesystem::path GetAppTestDataFilePath(const std::string& name);
 
 // Reads a test file from media/test/data directory and returns its content.
-std::vector<uint8_t> ReadTestDataFile(const std::string& name) {
-  std::filesystem::path path = GetTestDataFilePath(name);
-
-  std::vector<uint8_t> data;
-  data.resize(std::filesystem::file_size(path));
-
-  FILE* f = fopen(path.string().c_str(), "rb");
-  if (!f) {
-    LOG(FATAL) << "Failed to read test data from " << path;
-    return std::vector<uint8_t>();
-  }
-
-  size_t size = fread(data.data(), 1, data.size(), f);
-  data.resize(size);
-  fclose(f);
-
-  return data;
-}
+std::vector<uint8_t> ReadTestDataFile(const std::string& name);
 
 }  // namespace media
 }  // namespace shaka
